requirements.txt
django>=4.2
djangorestframework>=3.14.0
google-api-python-client>=2.90.0
google-auth>=2.20.0
requests>=2.28.0
python-dotenv>=0.21.0
